+++ kotlin-compiler-extensions/test-data/diagnostics/simple-1.kt
@@ -1,0 +1,6 @@
+package com.javiersc.kotlin.compiler.extensions.playground
+
+fun simple(): String {
+    val hello: String = run { "Hello" }
+    return hello
+}