+++ kotlin-compiler-extensions/test-data/diagnostics/simple-error-1.kt
@@ -1,0 +1,6 @@
+package com.javiersc.kotlin.compiler.extensions.playground
+
+fun simpleError(): String {
+    val hello: Int <!INITIALIZER_TYPE_MISMATCH!>=<!> "Hello"
+    return hello.toString()
+}