+++ kotlin-compiler-test-extensions/main/kotlin/com/javiersc/kotlin/compiler/test/services/ExtensionRegistrarConfigurator.kt
@@ -2,11 +2,39 @@ package com.javiersc.kotlin.compiler.test.services
 
 import org.jetbrains.kotlin.compiler.plugin.CompilerPluginRegistrar.ExtensionStorage
 import org.jetbrains.kotlin.config.CompilerConfiguration
+import org.jetbrains.kotlin.test.Constructor
+import org.jetbrains.kotlin.test.builders.TestConfigurationBuilder
 import org.jetbrains.kotlin.test.model.TestModule
+import org.jetbrains.kotlin.test.services.AbstractEnvironmentConfigurator
+import org.jetbrains.kotlin.test.services.AdditionalSourceProvider
 import org.jetbrains.kotlin.test.services.EnvironmentConfigurator
 import org.jetbrains.kotlin.test.services.TestServices
 
-internal class ExtensionRegistrarConfigurator(
+internal fun TestConfigurationBuilder.configurePlugin(
+    classpathProvider: Constructor<MetaRuntimeClasspathProvider>?,
+    additionalFilesProvider: Constructor<AdditionalFilesProvider>?,
+    registerCompilerExtensions: ExtensionStorage.(TestModule, CompilerConfiguration) -> Unit,
+) {
+    val configurators: List<Constructor<AbstractEnvironmentConfigurator>> = buildList {
+        add { testServices: TestServices ->
+            ExtensionRegistrarConfigurator(testServices, registerCompilerExtensions)
+        }
+        if (classpathProvider != null) {
+            useCustomRuntimeClasspathProviders(classpathProvider)
+            add { testServices: TestServices ->
+                classpathProvider(testServices).classpathConfigurator
+            }
+        }
+    }
+    useConfigurators(*configurators.toTypedArray())
+
+    val filesProvider: Constructor<AdditionalSourceProvider> = { testServices: TestServices ->
+        additionalFilesProvider?.invoke(testServices) ?: AdditionalFilesProvider(testServices)
+    }
+    useAdditionalSourceProviders(filesProvider)
+}
+
+private class ExtensionRegistrarConfigurator(
     testServices: TestServices,
     private val registerCompilerExtensions:
         ExtensionStorage.(TestModule, CompilerConfiguration) -> Unit,
@@ -16,6 +44,6 @@ internal class ExtensionRegistrarConfigurator(
         module: TestModule,
         configuration: CompilerConfiguration,
     ) {
-        this@ExtensionRegistrarConfigurator.registerCompilerExtensions(this, module, configuration)
+        registerCompilerExtensions(this, module, configuration)
     }
 }